requirements.txt
# ParkTegn dependencies
# Pillow is used for stitching satellite map tiles
Pillow
